requirements.txt
fastapi
uvicorn
python-dotenv
python-jose
passlib[bcrypt]
sqlalchemy
pymysql
email-validator
python-multipart
